+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Head
+
+- Add `useCapture` argument to `createTapListener`.
+
 ## 0.1.0
 
 - Initial release.